requirements.txt
aioboto3==13.2.0
aiohttp==3.9.5
diffusers==0.32.0.dev0
fastapi==0.115.6
graspologic==3.4.1
huggingface_hub==0.26.1
jsonlines==4.0.0
langchain_core==0.3.22
langchain_nvidia_ai_endpoints==0.3.5
lmdeploy==0.6.3
nano_vectordb==0.0.4.1
neo4j==5.26.0
nest_asyncio==1.6.0
networkx==3.3
numpy==2.1.3
ollama==0.4.4
openai==1.57.0
oracledb==2.5.0
pandas==2.2.3
peft==0.13.2
Pillow==11.0.0
pydantic==2.10.3
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
pyvis==0.3.2
setuptools==68.0.0
streamlit==1.39.0
tenacity==8.2.3
textract==1.6.5
tiktoken==0.8.0
torch==2.1.1
tqdm==4.66.4
transformers==4.41.2
uvicorn==0.32.1
